+++ R/module_functions.R
@@ -34,8 +34,8 @@ createModule <- function(constraints, item_pool, item_attrib, passage_attrib) {
 #' \code{\link{loadModules}} is a function for creating multiple \code{\linkS4class{module}} objects
 #' from a specification sheet.
 #'
-#' @param fn the name of a csv file containing module specifications.
-#' @param base_path (optional) the base path to append before the file paths contained in module specs.
+#' @param fn the full file path and name of a csv file containing module specifications.
+#' @param base_path (optional) the base path to be prepended to the file paths contained in the module specifications sheet.
 #' @param assessment_structure an \code{\linkS4class{assessment_structure}} object.
 #' @param examinee_list an examinee list from \code{\link{simExaminees}}. Used to determine the range of required modules.
 #'
@@ -51,7 +51,7 @@ createModule <- function(constraints, item_pool, item_attrib, passage_attrib) {
 #' }
 #'
 #' @returns a module list containing \code{\linkS4class{module}} objects.
-#' Each module can be accessed using \code{module_list[[grade]][[phase]]}.
+#' Each module can be accessed using \code{module_list[[grade]][[test]][[phase]]}.
 #'
 #' @examples
 #' assessment_structure <- createAssessmentStructure(
@@ -68,7 +68,7 @@ createModule <- function(constraints, item_pool, item_attrib, passage_attrib) {
 #'   assessment_structure = assessment_structure
 #' )
 #'
-#' fn <- system.file("extdata", "module_definition_MATH_normal_N500.csv", package = "maat")
+#' fn <- system.file("extdata", "module_definition_MATH_normal_N500_flexible.csv", package = "maat")
 #' pkg_path <- system.file(package = "maat")
 #' module_list <- loadModules(
 #'   fn,
@@ -99,32 +99,65 @@ loadModules <- function(fn, base_path = NULL, assessment_structure, examinee_lis
     }
   )
   required_grades <- unique(unlist(required_grades))
-
+  required_tests  <- sprintf("T%s", 1:assessment_structure@n_test)
   required_phases <- sprintf("P%s", 1:assessment_structure@n_phase)
-
   required_modules <- expand.grid(
     grade = required_grades,
+    test  = required_tests,
     phase = required_phases,
     stringsAsFactors = FALSE
   )
+  required_modules <- required_modules[
+    order(
+      required_modules$grade,
+      required_modules$test,
+      required_modules$phase
+    ),
+  ]
   n_required_modules <- dim(required_modules)[1]
 
   cat(sprintf("Required modules: %s\n", n_required_modules))
 
-  # Read module sheet and validate whether all required modules exist
+  # Read module sheet
   df <- read.csv(fn, stringsAsFactors = FALSE)
 
+  # Expand
+  if (!"Test" %in% colnames(df)) {
+    df <- cbind(Test = "", df)
+  }
+  df_list <- list()
+  for (i in 1:dim(df)[1]) {
+    df_row <- df[i, ]
+    if (df_row$Test == "") {
+      df_row <- df[rep(i, assessment_structure@n_test), ]
+      df_row$Test <- sprintf("T%s", 1:assessment_structure@n_test)
+    }
+    df_list[[i]] <- df_row
+  }
+  df <- do.call(rbind, df_list)
+  df <- df[order(
+    df$Grade, df$Test, df$Phase
+  ), ]
+
+  idx <- c("Grade", "Test", "Phase")
+  idx <- c(idx, setdiff(colnames(df), idx))
+  df <- df[, idx]
+  rownames(df) <- NULL
+
+  # Validate whether all required modules exist
   for (i in 1:n_required_modules) {
     idx <- which(
       required_modules$grade[i] == df$Grade &
+      required_modules$test[i] == df$Test &
       required_modules$phase[i] == df$Phase
     )
     if (length(idx) != 1) {
       stop(
         sprintf(
-          "cannot find Grade %s Phase %s in %s",
+          "cannot find Grade %s Test %s Phase %s in %s",
           required_modules$grade[i],
-          required_modules$phase[i],
+          required_modules$test[i],
+          required_modules$stage[i],
           fn
         )
       )
@@ -142,12 +175,14 @@ loadModules <- function(fn, base_path = NULL, assessment_structure, examinee_lis
 
     idx <- which(
       required_modules$grade[i] == df$Grade &
+      required_modules$test[i]  == df$Test  &
       required_modules$phase[i] == df$Phase
     )
 
     cat(sprintf(
-      "Grade %s Phase %s : Module %s\n",
+      "Grade %s Test %s Phase %s : Module %s\n",
       df$Grade[idx],
+      df$Test[idx],
       df$Phase[idx],
       df$Module[idx]
     ))
@@ -173,7 +208,10 @@ loadModules <- function(fn, base_path = NULL, assessment_structure, examinee_lis
     if (!df$Grade[idx] %in% names(module_list)) {
       module_list[[df$Grade[idx]]] <- list()
     }
-    module_list[[df$Grade[idx]]][[df$Phase[idx]]] <- o
+    if (!df$Test[idx] %in% names(module_list[[df$Grade[idx]]])) {
+      module_list[[df$Grade[idx]]][[df$Test[idx]]] <- list()
+    }
+    module_list[[df$Grade[idx]]][[df$Test[idx]]][[df$Phase[idx]]] <- o
 
   }