+++ inst/extdata/datasets.r
@@ -23,7 +23,7 @@ examinee_list_math <- simExaminees(
 )
 
 module_list_math <- loadModules(
-  "inst/extdata/module_definition_MATH_normal_N500.csv",
+  "inst/extdata/module_definition_MATH_normal_N500_flexible.csv",
   base_path = "inst",
   assessment_structure = assessment_structure_math,
   examinee_list = examinee_list_math